% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{emajor_example}
\alias{emajor_example}
\title{generate example Rmd files}
\usage{
emajor_example(exampleName)
}
\arguments{
\item{exampleName}{A character.}
}
\value{

}
\description{
generate example Rmd files
}
\examples{
none
}
