% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{emajor_grid_composite}
\alias{emajor_grid_composite}
\title{產生雙格成品}
\usage{
emajor_grid_composite(modifiedPlot1, modifiedPlot2, sources, png_filename)
}
\arguments{
\item{modifiedPlot1}{A list of ggplot object modified by emajor_theme()}

\item{modifiedPlot2}{A list of ggplot object modified by emajor_theme()}

\item{sources}{A character}

\item{png_filename}{A character}
}
\value{

}
\description{
產生雙格成品
}
\examples{
none
}
