% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{emajor_grid_single}
\alias{emajor_grid_single}
\title{產生單格成品}
\usage{
emajor_grid_single(modifiedPlot, sources, png_filename)
}
\arguments{
\item{modifiedPlot}{A list of ggplot object modified by emajor_theme()}

\item{sources}{A character}

\item{png_filename}{A character}
}
\value{

}
\description{
產生單格成品
}
\examples{
none
}
