% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{emajor_theme}
\alias{emajor_theme}
\title{Title}
\usage{
emajor_theme(baseplot, bar_col)
}
\arguments{
\item{baseplot}{A list, 底圖(ggplot 物件)}

\item{bar_col}{An integer. 是否為geom_bar 或 geom_col物件，是則1;否則0}
}
\value{

}
\description{
Title
}
