% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{emajor_themeBar}
\alias{emajor_themeBar}
\title{Theme for geom_bar}
\usage{
emajor_themeBar()
}
\value{

}
\description{
Theme for geom_bar
}
\examples{
ggplot(data=myData)+geom_bar(aes(x=x))+emajor_themeBar()
}
