% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{emajor_themeCol}
\alias{emajor_themeCol}
\title{Theme for geom_col}
\usage{
emajor_themeCol()
}
\value{

}
\description{
Theme for geom_col
}
\examples{
ggplot(data=myData)+geom_col(aes(x=x,y=y))+emajor_themeCol()
}
